+++ tests/ltypes1.py
@@ -1,11 +1,13 @@
+from ltypes import i8, i16, i32, i64
+
 def test_i8():
     i: i8
-    i = 5
+    i = i8(5)
     print(i)
 
 def test_i16():
     i: i16
-    i = 4
+    i = i16(4)
     print(i)
 
 def test_i32():
@@ -15,5 +17,5 @@ def test_i32():
 
 def test_i64():
     i: i64
-    i = 2
+    i = i64(2)
     print(i)
